requirements.txt
botbuilder-core>=4.14.0
aiohttp>=3.8.0
gunicorn>=20.1.0